Fixed C# syntax issue

+++ Scripts/player_script.cs
@@ -175,7 +175,7 @@ public partial class player_script : CharacterBody2D
     {
         if (area.Name == "Enemy_hitbox_container")
         {
-            hit(30*Upgrades.PlayerDamageMultiplier, area);
+            hit(30*(int)Upgrades.PlayerDamageMultiplier, area);
 
         }
     }
